Tratamento de erro

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -59,25 +60,44 @@ public class TopicosController {
 	}
 
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDTO detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.getById(id);
-		return new DetalhesDoTopicoDTO(topico);
+	public ResponseEntity<DetalhesDoTopicoDTO> detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+
+		if (topico.isPresent()) {
+			return ResponseEntity.ok(new DetalhesDoTopicoDTO(topico.get()));
+		}
+
+		return ResponseEntity.notFound().build();
+
 	}
 
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDTO> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
-		Topico topico = form.atualizar(id, topicoRepository);
+		Optional<Topico> optional = topicoRepository.findById(id);
 
-		return ResponseEntity.ok(new TopicoDTO(topico));
+		if (optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+
+			return ResponseEntity.ok(new TopicoDTO(topico));
+		}
+
+		return ResponseEntity.notFound().build();
+
 	}
 
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id) {
-		topicoRepository.deleteById(id);
+		Optional<Topico> optional = topicoRepository.findById(id);
 
-		return ResponseEntity.ok().build();
+		if (optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+
+		return ResponseEntity.notFound().build();
+
 	}
 
 }
